Add installation steps for Composer
Added installation steps for Composer using the global installation of PHP and the MAMP installation of PHP. 

Note that the use of HTML instead of standard Markdown was due to Markdown incorrectly handling nested ordered lists. This wasn't my preference, but I felt having consistent and clear styling of sub menus overrode ease of use as plain-text.